Fixed a bug where pulling the owner/group/others permissions from the FileMode raw value did not work properly, moved C interoperability code to new COptions folder and separated constructs to their own files
